added user-service.interface.ts

+++ server/src/app.controller.ts
@@ -33,11 +33,11 @@ export class AppController {
   @Get('/me')
   async getUserProfile(@Request() req ){
     try {
-      const user = await this.userService.findUser( req.user.userId );
+      const user = await this.userService.findUser( {id:req.user.userId} );
       const { _id, name, email, ChatGroups } = user;
-      const friendsData = await this.userService.getUsersFriendsData(req.user.userId);
+      const friendsData = await this.userService.getUsersFriendsData({userId:req.user.userId});
       const chatGroupDetails = await this.chatGroupService.getChatGroupDetails(ChatGroups);
-      const userProfileInfo = await this.userService.getUserProfileInfo(_id,name,email,chatGroupDetails,friendsData);
+      const userProfileInfo = await this.userService.getUserProfileInfo({id:_id, name:name, email:email, chatGroupDetails:chatGroupDetails, friendsData:friendsData});
       
       return userProfileInfo;
     } catch (error) {
@@ -48,9 +48,9 @@ export class AppController {
   @Post('/add-friends-to-chat-group/:chatGroupId/:friendId')
   async addFriendsToChatGroup(@Param('chatGroupId') chatGroupId: mongoose.Types.ObjectId ,@Param('friendId') friendId: mongoose.Types.ObjectId ){
     try {
-      const friendToAdd = await this.userService.findUser( friendId );
+      const friendToAdd = await this.userService.findUser( {id:friendId} );
       const updatedChatGroup = await this.chatGroupService.addUserToChatGroup( chatGroupId, friendToAdd );
-      await this.userService.addChatGroupToUser( friendId, updatedChatGroup );
+      await this.userService.addChatGroupToUser( {userId:friendId, chatGroup:updatedChatGroup} );
       return updatedChatGroup ;
     } catch (error) {
       throw new Error(error);
@@ -60,9 +60,9 @@ export class AppController {
   @Post('/remove-friends-from-chat-group/:chatGroupId/:friendId')
   async removeFriendsFromChatGroup(@Param('chatGroupId') chatGroupId: mongoose.Types.ObjectId, @Param('friendId') friendId: mongoose.Types.ObjectId ){
     try {
-      const friendToRemove = await this.userService.findUser( friendId );
+      const friendToRemove = await this.userService.findUser( {id:friendId} );
       const updatedChatGroup = await this.chatGroupService.removeUserFromChatGroup( chatGroupId, friendId );
-      await this.userService.removeChatGroupFromUser(friendToRemove,updatedChatGroup);
+      await this.userService.removeChatGroupFromUser({user: friendToRemove,chatGroup: updatedChatGroup});
       return updatedChatGroup;
     } catch (error) {
       throw new Error(error);
@@ -74,8 +74,8 @@ export class AppController {
       if( !req.user ) {
         throw new UnauthorizedException('You must be logged in for adding friend');
       }
-      const friend = await this.userService.findUser( friendId );
-      const updatedUser = await this.userService.addFriend( req.user.userId, friend );
+      const friend = await this.userService.findUser( {id:friendId} );
+      const updatedUser = await this.userService.addFriend( {userId: req.user.userId, friend: friend} );
       return updatedUser;
   }
   @UseGuards( JwtAuthGuard )
@@ -84,8 +84,8 @@ export class AppController {
       if( !req.user ) {
         throw new UnauthorizedException('You must be logged in for removing friend');
       }
-      const friend = await this.userService.findUser( friendId );
-      const updatedUser = await this.userService.removeFriend( req.user.userId, friend._id );
+      const friend = await this.userService.findUser( {id:friendId} );
+      const updatedUser = await this.userService.removeFriend( {userId: req.user.userId,friendId: friend._id });
       return updatedUser;
   }
   @UseGuards( JwtAuthGuard )
@@ -94,10 +94,10 @@ export class AppController {
     if (!req.user) {
       throw new UnauthorizedException('You must be logged in to view friends');
     }
-    const friends = await this.userService.getFriendsOfUser(req.user.userId);
+    const friends = await this.userService.getFriendsOfUser({userId:req.user.userId});
 
     const friendsData = await Promise.all(friends.map(async (friend) => {
-      const friendData = await this.userService.getUserData(friend);
+      const friendData = await this.userService.getUserData({userObject:friend});
       return friendData;
     }));
 
@@ -112,7 +112,7 @@ export class AppController {
     const friends = await this.chatGroupService.getChatGroupsUsers(chatGroupId);
 
     const friendsData = await Promise.all(friends.map(async (friend) => {
-      const friendData = await this.userService.getUserData(friend);
+      const friendData = await this.userService.getUserData({userObject: friend});
       return friendData;
     }));
     return friendsData;
@@ -125,7 +125,7 @@ export class AppController {
       }
       const user = await this.userService.findUser( req.user.userId );
       const newChatGroup = await this.chatGroupService.createChatGroup( body, user );
-      await this.userService.addChatGroupToUser( req.user.userId, newChatGroup );
+      await this.userService.addChatGroupToUser( {userId:req.user.userId,chatGroup: newChatGroup });
       return newChatGroup;
   }
   @UseGuards( JwtAuthGuard )
@@ -135,7 +135,7 @@ export class AppController {
       const chatGroupToBeDelete = await this.chatGroupService.getChatGroup( chatGroupId );
       const usersOfChatGroup = await this.chatGroupService.getChatGroupsUsers(chatGroupId);
       const deleteChatGroupFromUsers = usersOfChatGroup.map(async (user) => {
-        await this.userService.removeChatGroupFromUser(user, chatGroupToBeDelete);
+        await this.userService.removeChatGroupFromUser({user:user,chatGroup: chatGroupToBeDelete});
       });
       await Promise.all(deleteChatGroupFromUsers);
       await this.chatGroupService.deleteChatGroup(chatGroupId);
@@ -155,7 +155,7 @@ export class AppController {
   @UseGuards( JwtAuthGuard )
   @Get('/search-user')
   async searchUser( @Query("searchText") searchText: string ) {
-    const users = await this.userService.searchUser( searchText );
+    const users = await this.userService.searchUser( {searchText: searchText} );
     return users;
   }
   @UseGuards(JwtAuthGuard)
